wip fix hang on successful dev/bots/analyze.dart
TODO file issue; link PR that added timer

When running dev/bots/analyze.dart locally, if the analysis passes
then after printing "Analysis successful." the command hangs for a
long time.

It turns out this is due to the timer introduced recently as part of
cleaning up the output; on success, we'd wait 10 minutes for the
timeout to expire.  This doesn't affect CI because this feature
doesn't apply in CI (as detected by lack of color on stdout.)

Fix the issue by cleaning up the timer on success in the same way
as on failure.

While here, clean up the final summary messages slightly,
and also cut the trailing space that printProgress was leaving
on each line.