wip try to clarify how to set minSdkVersion and targetSdkVersion
Many developers have gotten the impression that the thing they have to
do in order to set these is either to put them in local.properties or
to edit their Flutter install.  The first is a very bad strategy: that
file is for local information, as its name says, and it shouldn't be
version-controlled and shared.  For anyone who wasn't intending to
start maintaining a local fork of Flutter, the second is an equally
bad strategy for much the same reason.

In particular, see most of the answers on this Stack Overflow question:
  https://stackoverflow.com/questions/52060516/flutter-how-to-change-android-minsdkversion-in-flutter-project
including some with hundreds of upvotes.

The intended, much simpler and more robust, approach is to simply edit
the lines that set them in the build script, like in any other Android
app.  In 99550 a year ago, some comments were added here to try to
clarify that, but people continue to be confused much more recently.
For example:
  https://stackoverflow.com/questions/52060516/flutter-how-to-change-android-minsdkversion-in-flutter-project#comment132078058_70316521
  https://stackoverflow.com/questions/52060516/flutter-how-to-change-android-minsdkversion-in-flutter-project#comment125321782_52060517
and issue 98442.

So try to make the comment more explicit: clearly indicate it's about
these two settings and not versionCode and versionName below, and give
a bit more context.

Also explain what the versionCode and versionName lines are doing, in
hopes of further clarifying the story.  The use of local.properties
for those is unconventional, and probably contributes to people
getting the idea they should do the same for the minSdkVersion and
targetSdkVersion.  The reason using that file for these is a good and
useful strategy, instead of a terrible one, is because of the help of
the `flutter` tool behind the scenes; so mention that explicitly, to
discourage over-generalizing from their example.