Clarify ChangeNotifier.dispose doc about the state it leaves behind
This wasn't quite accurate as written: calls to addListener after
disposal don't throw in general, only in debug mode.

The fact that removeListener is an exception to the rule is also
helpful information, worth making explicit here.

Also adjust the summary paragraph to fit the style guide, and to
match this class's other methods by being in the imperative.