wip clean up tree-walking in EditableText on scroll, using ScrollContext.notificationContext
Noticed this tree-walking while reading 138313 which introduced it,
and was curious if there was a better way and it looks like there is.

However: this part of the PR's logic is not actually tested at all!
Commenting out the calls to both these helper methods, so that we
proceed to handle the notification even if it's internal or otherwise
not from an ancestor, breaks no tests.

So although I believe this change is correct, there's a risk it's
wrong somehow, and it'd be best to have tests for it.  But given that
those tests ought to have been part of the original PR too, that's the
point where I lose motivation to make this cleanup.  If I thought it
made a big practical difference, I'd do it anyway; but even in this
version we're still doing *most* of the same tree-walking, just saving
the O(1) walk up from the

... hmm, maybe we can do better there too, with Scrollable.maybeOf.