and use Scrollable.maybeOf
Still not thrilled about writing the test, though.

Another thing this causes me to notice is that Scrollable.of/maybeOf
create a dependency, but few if any of their call sites actually
want that -- they're generally in gesture handlers, not build methods.
Could be useful to have non-dependency variants of those methods.
Not sure if there's any existing convention for that, though.