wip/ref Simplify _shouldMergeDirtyNodes logic in flushLayout
This _shouldMergeDirtyNodes flag should already be false here
thanks to the `while` loop above, so no need to clear it again.

(An exception can't be caused here by user layout code, because
_layoutWithoutResize already catches exceptions from performLayout.
So an uncaught exception here would have to be from a bug in the
framework; and it's impossible to systematically recover from those.)

As a bonus, this lets this `finally` block get elided completely in
release builds.  Compare 5759, though this isn't such a hot path
as to make the optimization important.