_isRelayoutBoundary step 3: Propagate "only" on tristate change
This relaxes the conditions under which [layout] causes a walk
of the tree to update or clear relayout boundaries.  Previously
we did so when _relayoutBoundary would change; now we do so
only when its tristate counterpart _isRelayoutBoundary would change.

This is a pure refactor because, in fact, the old and new conditions
are equivalent: the only times when _relayoutBoundary would change
are times when _isRelayoutBoundary would also change (and vice versa).

That's true because of the invariant documented on _relayoutBoundary,
namely that that property's value is either null, `this`, or
`parent!.relayoutBoundary!`.  The relayoutBoundary local, from its
definition, is also either `this` or `parent!.relayoutBoundary!`.
As a result, the values relayoutBoundary and _relayoutBoundary
will differ only when isRelayoutBoundary and _isRelayoutBoundary
also differ.