_isRelayoutBoundary step 4: Cut _relayoutBoundary
At this stage, the only places that ever read _relayoutBoundary are
either in the _isRelayoutBoundary getter, or purely in order to copy
the value to another object's _relayoutBoundary.  That means the only
information we ever ultimately use from _relayoutBoundary is the
information that's contained in the simpler _isRelayoutBoundary.

As a result, we can make _isRelayoutBoundary a field to store that
information directly, and cut out _relayoutBoundary entirely.