wip Skip actively propagating relayout boundaries to subtree; TODO test? via toStringShort?
Each call to RenderObject.layout, when it decides to change this
render object's _isRelayoutBoundary flag, also recursively visits
its descendants to adjust the same flag there.  For this it uses
helpers _cleanChildRelayoutBoundary and _propagateRelayoutBoundaryToChild.

This step is no longer necessary.  Removing it simplifies the code,
and is a small but measurable performance win: in a microbenchmark
with a linear tree of 4000 nested objects, measured on a Pixel 8,
this reduces the total time to compute each frame by about 41%, or
in absolute terms a bit over 0.1ms per frame.

---

It's conceptually natural that these are now obsolete.  Originally
they set the _relayoutBoundary property, which pointed to an
ancestor, not the mere flag _isRelayoutBoundary.  The pointer's
value naturally depends on changes above this node in the tree,
but there's no particular reason the flag should.

Moreover, the two calls actually have opposite effects!  Both
helpers walk the same set of descendants, where the flag is already
either false or null; _cleanChildRelayoutBoundary sets them all to
null, while _propagateRelayoutBoundaryToChild sets them all to
false.  And the choice between the two depends only on a condition
that seems to have nothing to do with relayout boundaries --
namely, whether this render object itself needs to actually perform
layout this time.  That suggests that whether those nodes have the
flag set to null vs. false doesn't really matter.

More systematically, to verify that these calls aren't needed:

 * In release builds, the only line that behaves differently when
   _isRelayoutBoundary is null vs. false is in `attach`.  If the
   render object is being attached after `layout` was called,
   there must have been an intervening `dropChild` to remove it
   from the tree.  And we still call _cleanChildRelayoutBoundary
   in that method.

 * The remaining two methods that behave differently on null vs.
   false are _debugRelayoutBoundaryAlreadyMarkedNeedsLayout, which
   is an assertion helper of markNeedsLayout, and toStringShort.

   Dropping _cleanChildRelayoutBoundary, so that the flag remains
   false in cases where it would have become null, causes the
   assertion helper to proceed with its checks and toStringShort
   to proceed with printing "relayoutBoundary=upNN".  These are
   correct as long as the invariant in _isRelayoutBoundary's doc
   holds; and there's no case where _cleanChildRelayoutBoundary
   here is needed to preserve that invariant.

   Dropping _propagateRelayoutBoundaryToChild, conversely, causes
   the assertion helper to skip checks it would have made and
   toStringShort to skip printing "relayoutBoundary=upNN".  Given
   that we've lived for years with _cleanChildRelayoutBoundary
   causing the same situation in the opposite case, this seems OK.