Let DrivenScrollActivity subclasses customize handling of overscroll
This functionality is useful in a custom DrivenScrollActivity
subclass, just like it's useful in a BallisticScrollActivity subclass.

The naming and documentation are copied from BallisticScrollActivity,
to keep these two classes aligned.  (It might also be useful to take
much of what they have in common and factor that out to a common
subclass; but that'd be for another PR.)